requirements.txt
discord.py
patreon
pyyaml
termtables